fix: fix pinned status aria-label/blurhash
